backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
gcri
comet-memory
python-dotenv>=1.0.0
loguru>=0.7.0
openai>=1.0.0
anthropic>=0.40.0
google-genai>=1.0.0
